requirements.txt
asgiref==3.2.10
Django==3.0.4
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
